move utils out of components
